requirements.txt
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
mysql-connector-python==8.2.0
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4
pandas==2.1.4

